+++ neurons/deployment_layer/model_e79286b7faed5eb22eaf30d38078fd30629876d5f62d734baa09930d20eb4acd/settings.json
@@ -1,0 +1,161 @@
+{
+    "run_args": {
+        "tolerance": {
+            "val": 0.0,
+            "scale": 1.0
+        },
+        "input_scale": 19,
+        "param_scale": 19,
+        "scale_rebase_multiplier": 1,
+        "lookup_range": [
+            0,
+            1048584
+        ],
+        "logrows": 21,
+        "num_inner_cols": 2,
+        "variables": [
+            [
+                "batch_size",
+                1
+            ]
+        ],
+        "input_visibility": "Public",
+        "output_visibility": "Public",
+        "param_visibility": "Fixed",
+        "rebase_frac_zero_constants": false,
+        "check_mode": "UNSAFE",
+        "commitment": "KZG",
+        "decomp_base": 16384,
+        "decomp_legs": 2,
+        "bounded_log_lookup": false
+    },
+    "num_rows": 701193,
+    "total_assignments": 1402386,
+    "total_const_size": 12,
+    "total_dynamic_col_size": 0,
+    "max_dynamic_input_len": 0,
+    "num_dynamic_lookups": 0,
+    "num_shuffles": 0,
+    "total_shuffle_col_size": 0,
+    "model_instance_shapes": [
+        [
+            256
+        ],
+        [
+            256
+        ],
+        [
+            256
+        ],
+        [
+            256
+        ],
+        [
+            256
+        ],
+        [
+            256
+        ],
+        [
+            256
+        ],
+        [
+            256
+        ],
+        [
+            256
+        ],
+        [
+            256
+        ],
+        [
+            1
+        ],
+        [
+            1
+        ],
+        [
+            1
+        ],
+        [
+            1
+        ],
+        [
+            1
+        ],
+        [
+            1
+        ],
+        [
+            1
+        ],
+        [
+            1
+        ],
+        [
+            1
+        ],
+        [
+            256
+        ],
+        [
+            1
+        ]
+    ],
+    "model_output_scales": [
+        19,
+        0
+    ],
+    "model_input_scales": [
+        0,
+        0,
+        0,
+        19,
+        19,
+        0,
+        0,
+        0,
+        0,
+        0,
+        19,
+        19,
+        19,
+        19,
+        19,
+        19,
+        19,
+        19,
+        0
+    ],
+    "module_sizes": {
+        "polycommit": [],
+        "poseidon": [
+            0,
+            [
+                0
+            ]
+        ]
+    },
+    "required_lookups": [
+        {
+            "Pow": {
+                "scale": 524288.0,
+                "a": 1.5
+            }
+        }
+    ],
+    "required_range_checks": [
+        [
+            0,
+            16383
+        ],
+        [
+            -1,
+            1
+        ]
+    ],
+    "check_mode": "UNSAFE",
+    "version": "18.1.5",
+    "num_blinding_factors": null,
+    "timestamp": 1738370108851
+}
